app/requirements.txt
Flask==1.0.2
bleach==3.1.0
Flask_SocketIO==4.2.1
Flask_Login==0.4.1
gevent==1.4.0
gevent-websocket==0.10.1
pymongo==3.8.0
libsass==0.19.2
pillow==6.1.0